Split Smartcat parser and refine language detection
